pyproject.toml
[project]
name = "action-recognition"
version = "0.1.0"
description = ""
authors = [
    {name = "edpavlova_1",email = "edpavlova_1@edu.hse.ru"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas (>=2.2.3,<3.0.0)",
    "rarfile (>=4.2,<5.0)",
    "requests (>=2.32.3,<3.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "torch (>=2.7.0,<3.0.0)",
    "torchvision (>=0.22.0,<0.23.0)",
    "pytorch-lightning (>=2.5.1.post0,<3.0.0)",
    "torchinfo (>=1.8.0,<2.0.0)",
    "av (>=14.4.0,<15.0.0)",
    "pre-commit (>=4.2.0,<5.0.0)",
    "hydra-core (>=1.3.2,<2.0.0)",
    "mlflow (>=2.22.0,<3.0.0)",
    "onnx (>=1.18.0,<2.0.0)",
    "onnxruntime (>=1.22.0,<2.0.0)",
    "tensorrt (>=10.11.0.33,<11.0.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
